Introducing MorphParts: You are able to set up some kernel features
Here are some features that are included in this parts as follows:
- Headphone gain
- Microphone gain
- Speaker gain
- Earpiece gain
- Spectrum
- Audio high quality
- Backlight dimmer (refactor for 4.9 only)
- FPS info
- Torch brightness (refactor for 4.9 only)
- Vibration strength (refactor for 4.9 only)
- SELinux changer
- Touch boost
- MSM thermal
- LMK
- TCPCA
- CPU boost
- GPU boost
- Ambient display gesture
- USB fastcharging
- Touch gesture

Notes: Most of the features above need kernel side support. You can take a look my kernel commit history (if your kernel version is 4.9) and find out some commits that are related with each feature above.

This won't happen without:
- @ronaxdevil
- @SonalSingh18